+++ app/src/main/java/org/routine_work/notepad/fragment/DeleteNotesFragment.java
@@ -107,17 +107,14 @@ public class DeleteNotesFragment extends ListFragment
 		boolean result = true;
 		Log.v(LOG_TAG, "Hello");
 
-		switch (item.getItemId())
-		{
-			case R.id.delete_note_menuitem:
-				Log.d(LOG_TAG, "delete_note_menuitem");
-				deleteCheckedNotes();
-				getActivity().setResult(Activity.RESULT_OK);
-				getActivity().finish();
-				break;
-			default:
-				result = super.onOptionsItemSelected(item);
-		}
+        if (item.getItemId() == R.id.delete_note_menuitem) {
+            Log.d(LOG_TAG, "delete_note_menuitem");
+            deleteCheckedNotes();
+            getActivity().setResult(Activity.RESULT_OK);
+            getActivity().finish();
+        } else {
+            result = super.onOptionsItemSelected(item);
+        }
 
 		Log.v(LOG_TAG, "Bye");
 		return result;
